+++ CHANGES.md
@@ -6,6 +6,9 @@
 
 * Implement `FromRequest` extractor for `Either<A,B>`
 
+* Implement `ResponseError` for `SendError`
+
+
 ## [0.7.15] - 2018-12-05
 
 ### Changed